fix: ensure user is not created twice

+++ apps/backend/src/modules/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { ConflictException, Injectable } from "@nestjs/common";
 import { existsSync, mkdirSync, readdirSync, readFileSync, writeFileSync } from "fs";
 import path from "path";
 import { User } from "./entities";
@@ -14,8 +14,8 @@ export class UsersService {
     }
 
     public saveUser(user: User): void {
-        this.users.push(user);
         this.saveUserToFile(user);
+        this.users.push(user);
     }
 
     public getUsers(): User[] {
@@ -43,12 +43,16 @@ export class UsersService {
         }
 
         return users;
-
     }
 
     private saveUserToFile(user: User): void {
         const filename = `${user.id}.json`;
         const filePath = path.join(USERS_DIR, filename);
+
+        if (existsSync(filePath)) {
+            throw new ConflictException(`User with id '${user.id}' already exists.`);
+        }
+
         writeFileSync(filePath, JSON.stringify(user), "utf-8");
     }
 }
